Make BGP ASN fields 64bit int

+++ pkg/agent/bgp/interface.go
@@ -94,7 +94,7 @@ type PeerConfig struct {
 type PeerStatus struct {
 	Address                    string
 	Port                       int32
-	ASN                        int32
+	ASN                        int64
 	MultihopTTL                int32
 	GracefulRestartTimeSeconds int32
 	SessionState               SessionState
